Ré-organisation des templatetags
