+++ previews/template_preview.py
@@ -142,7 +142,7 @@ class ComponentsPreview(LookbookPreview):
         }
         template = Template(
             """
-            {% load qfdmo_tags %}
+            {% load carte_tags %}
             {% acteur_pinpoint_tag acteur=acteur direction=direction action_list=action_list carte=carte carte_config=carte_config sous_categorie_id=sc_id %}
             """
         )
