+++ templates/data/_partials/contexte_details.html
@@ -1,4 +1,4 @@
-{% load qfdmo_tags %}
+{% load admin_data_tags %}
 
 <p>
     {% if identifiant_unique %}
