+++ templates/data/_partials/suppression_suggestion_details.html
@@ -1,6 +1,6 @@
 {% extends "data/_partials/suggestion_details.html" %}
 
-{% load qfdmo_tags %}
+{% load admin_data_tags %}
 
 {% block suggestion_title %}
     Acteur à supprimer
