+++ templates/data/_partials/value_details.html
@@ -1,4 +1,4 @@
-{% load qfdmo_tags %}
+{% load admin_data_tags %}
 
 {# Recursion on lists and dicts #}
 {% if value|valuetype == "list" and value %}
