+++ templates/layout/base.html
@@ -1,4 +1,4 @@
-{% load analytics_tags qfdmo_tags static %}
+{% load analytics_tags share_tags static %}
 <!DOCTYPE html>
 <html
     lang="fr"
