+++ templates/qfdmo/_address_card_partials/address_actions.html
@@ -1,4 +1,4 @@
-{% load qfdmo_tags %}<div class="qf-flex qf-justify-between">
+{% load carte_tags %}<div class="qf-flex qf-justify-between">
     <div class="fr-mb-1v">
         {% actions_for adresse direction as actions %}
         {% for action in actions %}
