+++ templates/qfdmo/acteur.html
@@ -1,5 +1,5 @@
 {% extends base_template %}
-{% load qfdmo_tags %}
+{% load carte_tags %}
 
 {% block page_title %}
 {{ object.libelle|title }} | {{ block.super }}
