+++ templates/qfdmo/acteur/_labels.html
@@ -1,4 +1,4 @@
-{% load qfdmo_tags %}
+{% load carte_tags %}
 <div class="qf-space-y-1w qf-pb-1w" data-testid="acteur-detail-labels">
     {% with label=object.labels_display|first %}
         {% if label.bonus %}
