+++ templates/qfdmo/carte.html
@@ -1,5 +1,5 @@
 {% extends base_template %}
-{% load qfdmo_tags static %}
+{% load static %}
 
 {% block content %}
 <turbo-frame
