+++ templates/qfdmo/carte/modals/filters.html
@@ -1,5 +1,6 @@
 {% extends 'qfdmo/_addresses_partials/filters/panel_filters.html' %}
-{% load qfdmo_tags %}
+
+{% load carte_tags %}
 
 {% block modal_title %}
 Filtres
