+++ templates/qfdmo/carte/pinpoints/acteur.html
@@ -1,4 +1,3 @@
-{% load qfdmo_tags %}
 <div class="qf-w-[34px] qf-h-[45px] {% if not force_visible %}qf-invisible{% endif %}"
      data-uuid="{{ acteur.uuid }}"
      data-latitude="{{ acteur.latitude }}"
