+++ templates/qfdmo/carte/results.html
@@ -6,7 +6,7 @@ The values of 120px and 150px are determined empirically.
 They need to match the css variables used for .grid-map selector in qfdmo.css
 {% endcomment %}
 
-{% load qfdmo_tags %}
+{% load carte_tags %}
 {% with address_ok=form.initial.adresse|default:form.initial.bounding_box|default:form.initial.epci_codes %}
 <div
     class="
