+++ templates/qfdmo/formulaire/results.html
@@ -5,7 +5,7 @@ The values of 120px and 150px are determined empirically.
 They need to match the css variables used for .grid-map selector in qfdmo.css
 {% endcomment %}
 
-{% load qfdmo_tags %}
+{% load carte_tags %}
 {% with address_ok=form.initial.adresse|default:form.initial.bounding_box|default:form.initial.epci_codes %}
 {% if request.user.is_authenticated and DEBUG %}
     <div class="qf-bg-blue-france-sun-113-625 qf-text-white qf-text-center qf-w-screen qf-py-1v qf-text-xs ">
