+++ templates/qfdmo/partials/action_list_by_direction.html
@@ -1,4 +1,4 @@
-{% load qfdmo_tags static %}
+{% load carte_tags static %}
 
 <div id="{{ action_direction }}_fieldset" data-search-solution-form-target='{{ action_direction }}'>
     <fieldset class="fr-fieldset fr-mx-0 fr-p-0 fr-mb-1w sm:qf-max-w-[600px]" aria-labelledby="action-by-direction-legend-{{ action_direction }}">
