+++ templates/qfdmo/partials/featureflip_inputs.html
@@ -1,4 +1,4 @@
-{% load qfdmo_tags static %}
+{% load share_tags static %}
 
 {% if is_iframe %}
     <input type="hidden" name="iframe">
